web-app/requirements.txt
Django>=2.0,<3.0
psycopg2
django-crispy-forms
Pillow
